requirements.txt
alabaster==0.7.10
anaconda-client==1.6.2
anaconda-navigator==1.5
anaconda-project==0.4.1
appnope==0.1.0
appscript==1.0.1
astroid==1.4.9
astropy==1.3
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.3
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.4
boto==2.46.1
Bottleneck==1.2.0
cffi==1.9.1
chardet==2.3.0
chest==0.2.3
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.7
configobj==5.0.6
contextlib2==0.5.4
cryptography==1.7.1
cvxopt==1.1.9
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dask==0.14.0
datashape==0.5.4
decorator==4.0.11
dill==0.2.5
docutils==0.13.1
entrypoints==0.2.2
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.12
Flask-Cors==3.0.2
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.7.1
h5py==2.6.0
HeapDict==1.0.0
html5lib==0.999
idna==2.2
imagesize==0.7.1
ipykernel==4.5.2
ipython==5.3.0
ipython-genutils==0.1.0
ipywidgets==6.0.0
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.9.0
Jinja2==2.9.5
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==5.0.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.0.0
lazy-object-proxy==1.2.2
llvmlite==0.16.0
locket==0.2.0
lxml==3.7.3
MarkupSafe==0.23
matplotlib==2.0.0
mistune==0.7.4
mpmath==0.19
multipledispatch==0.4.9
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
nltk==3.2.2
nose==1.3.7
notebook==4.4.1
numba==0.31.0
numexpr==2.6.2
numpy==1.12.0
numpydoc==0.6.0
odo==0.5.0
olefile==0.44
opencv-python==3.2.0.6
openpyxl==2.4.1
pandas==0.19.2
pandocfilters==1.4.1
partd==0.3.7
pathlib2==2.2.0
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
ply==3.10
prompt-toolkit==1.0.13
protobuf==3.2.0
psutil==5.2.0
ptyprocess==0.5.1
py==1.4.32
pyasn1==0.2.3
pycosat==0.6.1
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.5.0
Pygments==2.2.0
PyJWT==1.5.3
pylint==1.6.4
pyOpenSSL==16.2.0
pyparsing==2.1.4
PySocks==1.6.7
pytest==3.0.6
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.2.1
QtPy==1.2.1
redis==2.10.5
requests==2.13.0
rise==4.0.0b1
rope-py3k==0.9.4.post1
scikit-image==0.12.3
scikit-learn==0.19.1
scipy==0.19.0
seaborn==0.7.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.5.1
spyder==3.1.3
SQLAlchemy==1.1.6
statsmodels==0.8.0
sympy==1.0
tables==3.3.0
tensorflow==1.0.0
terminado==0.6
testpath==0.3
Theano==0.8.2
toolz==0.8.2
tornado==4.4.2
tqdm==4.11.2
traitlets==4.3.2
twilio==6.9.0
unicodecsv==0.14.1
wcwidth==0.1.7
Werkzeug==0.12
widgetsnbextension==2.0.0
wrapt==1.10.8
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.2
xlwt==1.2.0
